+++ test/library/middleware-test.ts
@@ -155,7 +155,7 @@ describe('middleware test', () => {
 	})
 
 	it('should let the first request through', async () => {
-		const app = createServer(rateLimit({ max: 1 }))
+		const app = createServer(rateLimit({ limit: 1 }))
 
 		await request(app).get('/').expect(200).expect('Hi there!')
 	})
@@ -163,7 +163,7 @@ describe('middleware test', () => {
 	it('should refuse additional connections once IP has reached the max', async () => {
 		const app = createServer(
 			rateLimit({
-				max: 2,
+				limit: 2,
 			}),
 		)
 
@@ -175,7 +175,7 @@ describe('middleware test', () => {
 	it('should (eventually) accept new connections from a blocked IP', async () => {
 		const app = createServer(
 			rateLimit({
-				max: 2,
+				limit: 2,
 				windowMs: 50,
 			}),
 		)
@@ -190,7 +190,7 @@ describe('middleware test', () => {
 	it('should work repeatedly', async () => {
 		const app = createServer(
 			rateLimit({
-				max: 2,
+				limit: 2,
 				windowMs: 50,
 			}),
 		)
@@ -206,12 +206,18 @@ describe('middleware test', () => {
 		await request(app).get('/').expect(200)
 	})
 
+	it('should block all requests if max is set to 0', async () => {
+		const app = createServer(rateLimit({ max: 0, validate: { limit: false } }))
+
+		await request(app).get('/').expect(429)
+	})
+
 	it('should show the provided message instead of the default message when max connections are reached', async () => {
 		const message = 'Enhance your calm'
 		const app = createServer(
 			rateLimit({
 				windowMs: 1000,
-				max: 2,
+				limit: 2,
 				message,
 			}),
 		)
@@ -225,7 +231,7 @@ describe('middleware test', () => {
 		const statusCode = 420
 		const app = createServer(
 			rateLimit({
-				max: 1,
+				limit: 1,
 				statusCode,
 			}),
 		)
@@ -243,7 +249,7 @@ describe('middleware test', () => {
 		const app = createServer(
 			rateLimit({
 				message,
-				max: 1,
+				limit: 1,
 			}),
 		)
 
@@ -255,7 +261,7 @@ describe('middleware test', () => {
 		const app = createServer(
 			rateLimit({
 				message: () => 'Too many requests.',
-				max: 1,
+				limit: 1,
 			}),
 		)
 
@@ -267,7 +273,7 @@ describe('middleware test', () => {
 		const app = createServer(
 			rateLimit({
 				message: async () => 'Too many requests.',
-				max: 1,
+				limit: 1,
 			}),
 		)
 
@@ -278,7 +284,7 @@ describe('middleware test', () => {
 	it('should use a custom handler when specified', async () => {
 		const app = createServer(
 			rateLimit({
-				max: 1,
+				limit: 1,
 				handler(_request, response) {
 					response.status(420).end('Enhance your calm')
 				},
@@ -292,7 +298,7 @@ describe('middleware test', () => {
 	it('should allow custom key generators', async () => {
 		const app = createServer(
 			rateLimit({
-				max: 2,
+				limit: 2,
 				keyGenerator: (request, _response) => request.query.key as string,
 			}),
 		)
@@ -311,7 +317,7 @@ describe('middleware test', () => {
 	it('should allow custom skip function', async () => {
 		const app = createServer(
 			rateLimit({
-				max: 2,
+				limit: 2,
 				skip: () => true,
 			}),
 		)
@@ -324,7 +330,7 @@ describe('middleware test', () => {
 
 	it('should allow custom skip function that returns a promise', async () => {
 		const limiter = rateLimit({
-			max: 2,
+			limit: 2,
 			skip: async () => true,
 		})
 
@@ -338,7 +344,7 @@ describe('middleware test', () => {
 	it('should allow max to be a function', async () => {
 		const app = createServer(
 			rateLimit({
-				max: () => 2,
+				limit: () => 2,
 			}),
 		)
 
@@ -350,7 +356,7 @@ describe('middleware test', () => {
 	it('should allow max to be a function that returns a promise', async () => {
 		const app = createServer(
 			rateLimit({
-				max: async () => 2,
+				limit: async () => 2,
 			}),
 		)
 
@@ -362,7 +368,7 @@ describe('middleware test', () => {
 	it('should calculate the remaining hits', async () => {
 		const app = createServer(
 			rateLimit({
-				max: async () => 2,
+				limit: async () => 2,
 			}),
 		)
 
@@ -674,7 +680,7 @@ describe('middleware test', () => {
 		async (name, store) => {
 			const app = createServer(
 				rateLimit({
-					max: 2,
+					limit: 2,
 					store,
 					skipFailedRequests: true,
 				}),
@@ -699,7 +705,7 @@ describe('middleware test', () => {
 
 			const app = createServer(
 				rateLimit({
-					max: 1,
+					limit: 1,
 					store,
 					handler() {
 						const exception = new Error('420: Enhance your calm')
@@ -738,7 +744,7 @@ describe('middleware test', () => {
 
 			const app = createServer(
 				rateLimit({
-					max: 1,
+					limit: 1,
 					store,
 					skip() {
 						const exception = new Error('420: Enhance your calm')
@@ -785,15 +791,25 @@ describe('middleware test', () => {
 		])
 
 		await request(app).get('/').expect(200)
-		expect(savedRequestObject).toBeTruthy()
-		expect(savedRequestObject.rateLimit).toBeTruthy()
-		expect(savedRequestObject.rateLimit.limit).toEqual(5)
-		expect(savedRequestObject.rateLimit.remaining).toEqual(4)
+		expect(savedRequestObject?.rateLimit).toMatchObject({
+			limit: 5,
+			used: 1,
+			remaining: 4,
+			resetTime: expect.any(Date),
+		})
+
+		// Make sure the hidden proerty is also set.
+		expect(savedRequestObject?.rateLimit.current).toBe(1)
 
 		savedRequestObject = undefined
 		await request(app).get('/').expect(200)
-		expect(savedRequestObject.rateLimit.limit).toEqual(5)
-		expect(savedRequestObject.rateLimit.remaining).toEqual(3)
+		expect(savedRequestObject?.rateLimit).toMatchObject({
+			limit: 5,
+			used: 2,
+			remaining: 3,
+			resetTime: expect.any(Date),
+		})
+		expect(savedRequestObject?.rateLimit.current).toBe(2)
 	})
 
 	it('should pass the number of hits and the limit to the next request handler with a custom property', async () => {
@@ -817,20 +833,28 @@ describe('middleware test', () => {
 		])
 
 		await request(app).get('/').expect(200)
-		expect(savedRequestObject).toBeTruthy()
-		expect(savedRequestObject.rateLimitInfo).toBeTruthy()
-		expect(savedRequestObject.rateLimitInfo.limit).toEqual(5)
-		expect(savedRequestObject.rateLimitInfo.remaining).toEqual(4)
+		expect(savedRequestObject?.rateLimitInfo).toMatchObject({
+			limit: 5,
+			used: 1,
+			remaining: 4,
+			resetTime: expect.any(Date),
+		})
+		expect(savedRequestObject?.rateLimitInfo.current).toBe(1)
 
 		savedRequestObject = undefined
 		await request(app).get('/').expect(200)
-		expect(savedRequestObject.rateLimitInfo.limit).toEqual(5)
-		expect(savedRequestObject.rateLimitInfo.remaining).toEqual(3)
+		expect(savedRequestObject?.rateLimitInfo).toMatchObject({
+			limit: 5,
+			used: 2,
+			remaining: 3,
+			resetTime: expect.any(Date),
+		})
+		expect(savedRequestObject?.rateLimitInfo.current).toBe(2)
 	})
 
 	it('should handle two rate-limiters with different `requestPropertyNames` operating independently', async () => {
 		const keyLimiter = rateLimit({
-			max: 2,
+			limit: 2,
 			requestPropertyName: 'rateLimitKey',
 			keyGenerator: (request) => request.query.key as string,
 			handler(_request, response) {
@@ -838,7 +862,7 @@ describe('middleware test', () => {
 			},
 		})
 		const globalLimiter = rateLimit({
-			max: 5,
+			limit: 5,
 			requestPropertyName: 'rateLimitGlobal',
 			keyGenerator: () => 'global',
 			handler(_request, response) {