+++ test/library/options-test.ts
@@ -8,6 +8,47 @@ import rateLimit, {
 } from '../../source/index.js'
 
 describe('options test', () => {
+	class MockStore implements Store {
+		options!: Options
+
+		init(options: Options): void {
+			this.options = options
+		}
+
+		async get(_key: string): Promise<ClientRateLimitInfo> {
+			return { totalHits: 1, resetTime: undefined }
+		}
+
+		async increment(_key: string): Promise<ClientRateLimitInfo> {
+			return { totalHits: 1, resetTime: undefined }
+		}
+
+		async decrement(_key: string): Promise<void> {}
+
+		async resetKey(_key: string): Promise<void> {}
+	}
+
+	// TODO: Update in v7.
+	it('should allow the use of pre-6.x headers options', async () => {
+		const store = new MockStore()
+		rateLimit({
+			store,
+			headers: false,
+		})
+
+		expect(store.options.headers).toEqual(false)
+	})
+
+	it('should allow the use of the `max` option', async () => {
+		const store = new MockStore()
+		rateLimit({
+			store,
+			max: 7,
+		})
+
+		expect(store.options.limit).toEqual(7)
+	})
+
 	it('should not allow the use of an invalid store', async () => {
 		class InvalidStore {
 			invalid = true
@@ -19,35 +60,6 @@ describe('options test', () => {
 				store: new InvalidStore(),
 			})
 		}).toThrowError(/store/)
-	})
-
-	// TODO: Update in v7.
-	it('should allow the use of pre-6.x headers options', async () => {
-		class MockStore implements Store {
-			options!: Options
-
-			init(options: Options): void {
-				this.options = options
-			}
-
-			async increment(_key: string): Promise<ClientRateLimitInfo> {
-				return { totalHits: 1, resetTime: undefined }
-			}
-
-			async decrement(_key: string): Promise<void> {}
-
-			async resetKey(_key: string): Promise<void> {}
-		}
-
-		const store = new MockStore()
-		rateLimit({
-			store,
-			headers: false,
-			draft_polli_ratelimit_headers: true, // eslint-disable-line @typescript-eslint/naming-convention
-		})
-
-		expect(store.options.headers).toEqual(false)
-		expect(store.options.draft_polli_ratelimit_headers).toEqual(true)
 	})
 
 	it('should not call `init` if it is not a function', async () => {